handle event refactor done
