+++ server/meep-draft/defaultControls/color/component.react.js
@@ -3,6 +3,19 @@
  */
 import React, { Component } from 'react';
 
+import Draft, {
+  Editor,
+  EditorState,
+  Modifier,
+  RichUtils,
+  convertToRaw,
+  convertFromRaw,
+  Entity,
+  CompositeDecorator,
+  ContentState,
+  SelectionState,
+  CharacterMetadata } from 'draft-js';
+
 import merge from '../../lib/merge.js';
 import styles from '../../draft-text.style';
 
@@ -10,7 +23,11 @@ import ColorButton from './color-button.react';
 
 import {
   COLORSTYLE
-} from '../../draft-type-core/inline'
+} from '../../draft-type-core/inline';
+
+import {
+  COLORS
+} from '../../draft-custom-core/custom';
 
 export default class ColorControls extends Component {
   constructor(props, context) {
@@ -26,6 +43,38 @@ export default class ColorControls extends Component {
       })
     }
   }
+
+  _toggleColor = (color) => {
+    const {
+      editorState,
+      onChange
+    } = this.props
+    const selection = editorState.getSelection();
+    //最多只能一次有一個顏色
+    const nextContentState = Object.keys(COLORS)
+                             .reduce((contentState, color) => {
+                               return Modifier.removeInlineStyle(contentState, selection, color)
+                             }, editorState.getCurrentContent());
+    let nextEditorState = EditorState.push(
+      editorState,
+      nextContentState,
+      'change-inline-style'
+    )
+    const currentStyle = editorState.getCurrentInlineStyle();
+    if(selection.isCollapsed()) {
+      nextEditorState = currentStyle.reduce((state, color) => {
+        return RichUtils.toggleInlineStyle(state, color);
+      }, nextEditorState);
+    }
+    if(!currentStyle.has(color)) {
+      nextEditorState = RichUtils.toggleInlineStyle(
+        nextEditorState,
+        color
+      )
+    }
+    onChange(nextEditorState)
+  }
+
   render() {
     let {
       editorState
@@ -38,7 +87,9 @@ export default class ColorControls extends Component {
           active={currentStyle.has(type.style)}
           label={type.label}
           style={type.style}
-          onToggle={this.props.onToggle}
+          onToggle={() => {
+            this._toggleColor(type.style)
+          }}
         />)
     })) : ( null )
     return (
