+++ server/meep-draft/defaultControls/font-size/component.react.js
@@ -3,6 +3,19 @@
  */
 import React, { Component } from 'react';
 
+import Draft, {
+  Editor,
+  EditorState,
+  Modifier,
+  RichUtils,
+  convertToRaw,
+  convertFromRaw,
+  Entity,
+  CompositeDecorator,
+  ContentState,
+  SelectionState,
+  CharacterMetadata } from 'draft-js';
+
 import merge from '../../lib/merge.js';
 import styles from '../../draft-text.style';
 
@@ -11,6 +24,11 @@ import SelectItem from './select-item.react';
 import {
   FONTSIZESTYLE
 } from '../../draft-type-core/inline'
+
+//custom-core
+import {
+  FONTSIZE
+} from '../../draft-custom-core/custom'
 
 export default class FontSizeControls extends Component {
   constructor(props, context) {
@@ -26,6 +44,38 @@ export default class FontSizeControls extends Component {
       })
     }
   }
+
+  _toggleFontSize = (size) => {
+    let {
+      editorState,
+      onChange
+    } = this.props
+    const selection = editorState.getSelection();
+    //最多只能一次有一個顏色
+    const nextContentState = Object.keys(FONTSIZE)
+                             .reduce((contentState, size) => {
+                               return Modifier.removeInlineStyle(contentState, selection, size)
+                             }, editorState.getCurrentContent());
+    let nextEditorState = EditorState.push(
+      editorState,
+      nextContentState,
+      'change-inline-style'
+    )
+    const currentStyle = editorState.getCurrentInlineStyle();
+    if(selection.isCollapsed()) {
+      nextEditorState = currentStyle.reduce((state, size) => {
+        return RichUtils.toggleInlineStyle(state, size);
+      }, nextEditorState);
+    }
+    if(!currentStyle.has(size)) {
+      nextEditorState = RichUtils.toggleInlineStyle(
+        nextEditorState,
+        size
+      )
+    }
+    onChange(nextEditorState)
+  }
+
   render() {
     let currentStyle = this.props.editorState.getCurrentInlineStyle();
     let fontSize = 10;
@@ -40,7 +90,7 @@ export default class FontSizeControls extends Component {
             key={`font_size_button_${idx}`}
             size={size}
             style={FONTSIZESTYLE[idx].style}
-            onToggle={this.props.onToggle}
+            onToggle={this._toggleFontSize}
             onOpen={this._onOpen}
           />
         )
