+++ server/meep-draft/draft-text.react.js
@@ -153,29 +153,8 @@ export default class DraftText extends Component {
       })
     }
     //
-    this.onDoHandle = (editorState, action) => {
-      let newEditorState
-      switch(action) {
-        case 'undo':
-          newEditorState = EditorState.undo(editorState);
-        break;
-        case 'redo':
-          newEditorState = EditorState.redo(editorState);
-        break;
-      }
-      if(newEditorState) {
-        this._setEditorState(newEditorState);
-      }
-    };
-    //
-    this.toggleFontSize = (size) => this._toggleFontSize(size);
-    this.toggleFontFamily = (family) => this._toggleFontFamily(family);
-    this.toggleInlineStyle = (style) => this._toggleInlineStyle(style);
-    this.toggleBlockType = (type) => this._toggleBlockType(type);
-    this.toggleColor = (color) => this._toggleColor(color);
-    this.toggleBackgroundColor = (backgroundcolor) => this._toggleBackgroundColor(backgroundcolor);
-    this.toggleAlign = (align) => this._toggleAlign(align);
-    //
+
+
     this.logState = () => {
       const content = this.state.editorState.getCurrentContent();
       console.log(convertToRaw(content))
@@ -183,221 +162,6 @@ export default class DraftText extends Component {
     this.logClear = () => {
       console.clear()
     }
-  }
-
-  _onHandlLink = (e, action) => {
-    const {
-      editorState
-    } = this.state;
-    const selection = editorState.getSelection();
-    let entityKey;
-    let content;
-    let oldUrl;
-    switch(action) {
-      case 'addLink':
-        if (selection.isCollapsed()) {
-          return
-        }
-        let selectedBlockEntityNumber = editorState
-                                .getCurrentContent()
-                                .getBlockForKey(editorState.getSelection().getStartKey())
-                                .getEntityAt(editorState.getSelection().getStartOffset());
-        if(selectedBlockEntityNumber !== null) {
-          oldUrl = Entity.get(selectedBlockEntityNumber).get('data').href;
-        }
-        const href = window.prompt('請輸入網址', oldUrl);
-        content = editorState.getCurrentContent();
-        if(href === null) return
-        entityKey = Entity.create('link', 'MUTABLE', {href});
-      break;
-      case 'removeLink':
-        entityKey = null;
-        content = editorState.getCurrentContent();
-      break;
-    }
-    this.setState({
-      editorState: RichUtils.toggleLink(editorState, selection, entityKey),
-    });
-  }
-
-  _setEditorState = (editorState) => {
-    this.onChange(editorState);
-  }
-
-  _toggleFontFamily = (family) => {
-    const {
-      editorState
-    } = this.state
-    const selection = editorState.getSelection();
-    //最多只能一次有一個顏色
-    const nextContentState = Object.keys(FONTFAMILY)
-                             .reduce((contentState, family) => {
-                               return Modifier.removeInlineStyle(contentState, selection, family)
-                             }, editorState.getCurrentContent());
-    let nextEditorState = EditorState.push(
-      editorState,
-      nextContentState,
-      'change-inline-style'
-    )
-    const currentStyle = editorState.getCurrentInlineStyle();
-    if(selection.isCollapsed()) {
-      nextEditorState = currentStyle.reduce((state, family) => {
-        return RichUtils.toggleInlineStyle(state, family);
-      }, nextEditorState);
-    }
-    if(!currentStyle.has(family)) {
-      nextEditorState = RichUtils.toggleInlineStyle(
-        nextEditorState,
-        family
-      )
-    }
-    this.onChange(nextEditorState)
-  }
-
-  _toggleFontSize = (size) => {
-    let {
-      editorState
-    } = this.state
-    const selection = editorState.getSelection();
-    //最多只能一次有一個顏色
-    const nextContentState = Object.keys(FONTSIZE)
-                             .reduce((contentState, size) => {
-                               return Modifier.removeInlineStyle(contentState, selection, size)
-                             }, editorState.getCurrentContent());
-    let nextEditorState = EditorState.push(
-      editorState,
-      nextContentState,
-      'change-inline-style'
-    )
-    const currentStyle = editorState.getCurrentInlineStyle();
-    if(selection.isCollapsed()) {
-      nextEditorState = currentStyle.reduce((state, size) => {
-        return RichUtils.toggleInlineStyle(state, size);
-      }, nextEditorState);
-    }
-    if(!currentStyle.has(size)) {
-      nextEditorState = RichUtils.toggleInlineStyle(
-        nextEditorState,
-        size
-      )
-    }
-    this.onChange(nextEditorState)
-  }
-
-  _toggleInlineStyle = (inlineStyle) => {
-    const {
-      editorState
-    } = this.state
-    this.onChange(
-      RichUtils.toggleInlineStyle(
-        editorState,
-        inlineStyle
-      )
-    );
-  }
-
-  _toggleBlockType = (blockType) => {
-    this.onChange(
-      RichUtils.toggleBlockType(
-        this.state.editorState,
-        blockType
-      )
-    );
-  }
-
-  _toggleColor = (color) => {
-    const {
-      editorState
-    } = this.state
-    const selection = editorState.getSelection();
-    //最多只能一次有一個顏色
-    const nextContentState = Object.keys(COLORS)
-                             .reduce((contentState, color) => {
-                               return Modifier.removeInlineStyle(contentState, selection, color)
-                             }, editorState.getCurrentContent());
-    let nextEditorState = EditorState.push(
-      editorState,
-      nextContentState,
-      'change-inline-style'
-    )
-    const currentStyle = editorState.getCurrentInlineStyle();
-    if(selection.isCollapsed()) {
-      nextEditorState = currentStyle.reduce((state, color) => {
-        return RichUtils.toggleInlineStyle(state, color);
-      }, nextEditorState);
-    }
-    if(!currentStyle.has(color)) {
-      nextEditorState = RichUtils.toggleInlineStyle(
-        nextEditorState,
-        color
-      )
-    }
-    this.onChange(nextEditorState)
-  }
-
-  _toggleBackgroundColor = (backgroundcolor) => {
-    const {
-      editorState
-    } = this.state
-    const selection = editorState.getSelection();
-    //最多只能一次有一個顏色
-    const nextContentState = Object.keys(BACKGROUNDCOLORS)
-                             .reduce((contentState, backgroundcolor) => {
-                               return Modifier.removeInlineStyle(contentState, selection, backgroundcolor)
-                             }, editorState.getCurrentContent());
-    let nextEditorState = EditorState.push(
-      editorState,
-      nextContentState,
-      'change-inline-style'
-    )
-    const currentStyle = editorState.getCurrentInlineStyle();
-    if(selection.isCollapsed()) {
-      nextEditorState = currentStyle.reduce((state, backgroundcolor) => {
-        return RichUtils.toggleInlineStyle(state, backgroundcolor);
-      }, nextEditorState);
-    }
-    if(!currentStyle.has(backgroundcolor)) {
-      nextEditorState = RichUtils.toggleInlineStyle(
-        nextEditorState,
-        backgroundcolor
-      )
-    }
-    this.onChange(nextEditorState)
-  }
-
-  _toggleAlign = (align) => {
-    const {
-      editorState
-    } = this.state
-    const selection = editorState.getSelection();
-    const nextContentState = Object.keys(ALIGN)
-                             .reduce((contentState, align) => {
-                               return Modifier.removeInlineStyle(contentState, selection, align)
-                             }, editorState.getCurrentContent());
-    let nextEditorState = EditorState.push(
-      editorState,
-      nextContentState,
-      'change-inline-style'
-    )
-    const currentStyle = editorState.getCurrentInlineStyle();
-    if(selection.isCollapsed()) {
-      nextEditorState = currentStyle.reduce((state, align) => {
-        return RichUtils.toggleInlineStyle(state, align);
-      }, nextEditorState);
-    }
-    if(!currentStyle.has(align)) {
-      nextEditorState = RichUtils.toggleInlineStyle(
-        nextEditorState,
-        align
-      )
-    }
-    this.onChange(nextEditorState)
-    this.onChange(
-      RichUtils.toggleBlockType (
-        this.state.editorState,
-        align
-      )
-    );
   }
 
   componentWillMount() {
@@ -436,6 +200,10 @@ export default class DraftText extends Component {
   render() {
     const {editorState} = this.state;
 
+    const {
+      controls
+    } = this.defaultSetting
+
     let StateLog = this.state.editMode ? (
       <span style={{fontSize: '14px'}}>
         {' '}
@@ -461,69 +229,66 @@ export default class DraftText extends Component {
     const rootStyle = this.checkRootStyle ? (this.props.editorStyle.root) : ({})
     const rootControlStyle = this.checkRootControlStyle ? (this.props.editorStyle['root-control']) : ({})
     const rootInputStyle = this.checkRootInputStyle ? (this.props.editorStyle['root-input']) : ({})
-
-    const {
-      controls
-    } = this.defaultSetting
-
+    
     const fontFamilyControls = controls.fontFamily ? (
       <FontFamilyControls
         editorState={editorState}
-        onToggle={this.toggleFontFamily}
         customStyle={rootControlStyle}
+        onChange={this.onChange}
       />
     ) : null
 
     const fontSizeControls = controls.fontSize ? (
       <FontSizeControls
         editorState={editorState}
-        onToggle={this.toggleFontSize}
         customStyle={rootControlStyle}
+        onChange={this.onChange}
       />
     ) : null
 
     const textControls = controls.text ? (
       <TextControls
         editorState={editorState}
-        onToggle={this.toggleInlineStyle}
         groupControls={controls.text}
+        onChange={this.onChange}
       />
     ) : null
 
     const linkControls = controls.link ? (
       <LinkControls
-        onHandlLink={this._onHandlLink}
+        editorState={editorState}
         groupControls={controls.link}
+        onChange={this.onChange}
       />
     ) : null
 
     const blockControls = controls.block ? (
       <BlockControls
         editorState={editorState}
-        onToggle={this.toggleBlockType}
         groupControls={controls.block}
+        onChange={this.onChange}
       />
     ) : null
 
     const colorControls = controls.color ? (
       <ColorControls
         editorState={editorState}
-        onToggle={this.toggleColor}
+        onChange={this.onChange}
       />
     ) : null
 
     const backgrounControls = controls.background ? (
       <BackgroundControls
         editorState={editorState}
-        onToggle={this.toggleBackgroundColor}
+        onChange={this.onChange}
       />
     ) : null
 
     const contentControls = controls.content ? (
       <ContentControls
         editorState={editorState}
-        onDoHandle={this.onDoHandle}
         groupControls={controls.content}
+        onChange={this.onChange}
       />
     ) : null
